feat: Implement order management features and enhance user account pages
- Added OrderController to handle order-related requests.
- Created OrderRepoImpl for order data persistence.
- Introduced OrderService for business logic related to orders.
- Developed JSP views for account and order history.
- Updated navigation links in home, login, product, and register pages.
- Enhanced user experience with dynamic account details update functionality.
- Implemented order cancellation feature with corresponding UI updates.
- Refactored CSS for account and order pages for improved styling.
- Added JavaScript for handling form submissions and dynamic UI updates.

+++ src/main/java/com/techouts/ecommerce/model/OrderItem.java
@@ -5,7 +5,6 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
-import java.math.BigDecimal;
 
 @Entity
 @Table(name = "order_items")
@@ -29,9 +28,9 @@ public class OrderItem {
     private int quantity;
 
     @Column(name = "purchased_price")
-    private BigDecimal purchasedPrice;
+    private float purchasedPrice;
 
-    public OrderItem(Product productId, Order orderId, int quantity, BigDecimal purchasedPrice) {
+    public OrderItem(Product productId, Order orderId, int quantity, float purchasedPrice) {
 
         this.productId = productId;
         this.orderId = orderId;
